[SPARK-46265][CONNECT] Assertions in AddArtifact RPC make the connect client incompatible with older clusters
### What changes were proposed in this pull request?

A previous commit - f061fc61 - updated the response of the
`AddArtifact` RPC to return the session id. Further, it also added
an assertion to the client of this RPC asserting that the session id
returned by the server matches the session id that was requested.

However, we will have the case that a connect client with this new
assertion included may connect to a Spark cluster that does not yet
have this change, making the change backwards incompatible.

Loosen the assertion to allow empty session id in the RPC's response.

### Why are the changes needed?

Newer connect clients can connect with older Spark clusters.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Manually started the a connect cluster using `spark-connect-shell`
and using the connect client shell via `spark-connect-scala-client`.
Changed the service to not return session id and verified the fix.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44181 from nija-at/artifact-sessionid.

Authored-by: Niranjan Jayakar <nija@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>