[SPARK-46322][PYTHON][DOCS] Replace external link with internal link for error documentation
### What changes were proposed in this pull request?

This PR proposes to update the documentation in `errors.rst` by replacing external hyperlink with internal (relative) link. Specifically, it changes the link in the 'Contributing Error and Exception' section from an absolute URL to a relative path.

### Why are the changes needed?

Because using internal links in the documentation ensures that the links remain valid and consistent with the latest documentation, even after updates or restructuring. It also reduces the risk of broken links in the future. Internal links provide a more reliable and maintainable approach to linking related content within the PySpark documentation.

### Does this PR introduce _any_ user-facing change?

No, it only affects the documentation's hyperlinks, improving their reliability and maintenance.

### How was this patch tested?

The changes were manually reviewed to ensure that the new internal links correctly point to the intended sections of the documentation.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44251 from itholic/SPARK-46322.

Authored-by: Haejoon Lee <haejoon.lee@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>