[MINOR][SQL] Convert `UnresolvedException` to an internal error
### What changes were proposed in this pull request?
In the PR, I propose to change the parent class of `UnresolvedException` from `AnalysisException` to `SparkException` with the error class `INTERNAL_ERROR`. If an user observes the error, this is definitely a bug in the compiler.

### Why are the changes needed?
To unify all Spark exceptions, and assign an error class. So, this should improve user experience with Spark SQL.

### Does this PR introduce _any_ user-facing change?
No, users shouldn't face to the error in regular cases.

### How was this patch tested?
By existing GAs.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #44311 from MaxGekk/error-class-UnresolvedException.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>