[SPARK-46381][SQL] Migrate sub-classes of `AnalysisException` to error classes
### What changes were proposed in this pull request?
In the PR, I propose to migrate the rest two sub-classes of `AnalysisException` onto error classes:
- NonEmptyNamespaceException
- ExtendedAnalysisException

and forbid raising of such exception without an error class.

### Why are the changes needed?
This is a part of the migration on the error framework, and unifying errors in Spark.

### Does this PR introduce _any_ user-facing change?
Yes, if user's code depends on the format of error messages.

### How was this patch tested?
By existing test suites like:
```
$ PYSPARK_PYTHON=python3 build/sbt "sql/testOnly org.apache.spark.sql.SQLQueryTestSuite"
```

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #44314 from MaxGekk/error-class-ExtendedAnalysisException.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>