[SPARK-46294][SQL] Clean up semantics of init vs zero value
### What changes were proposed in this pull request?

Cleaning up the semantics of init and zero value to the following. This also helps define what an "invalid" metric is.

initValue is the starting value for a SQLMetric. If a metric has value equal to its initValue, then it can/should be filtered out before aggregating with SQLMetrics.stringValue().

zeroValue defines the lowest value considered valid. If a SQLMetric is invalid, it is set to zeroValue upon receiving any updates, and it also reports zeroValue as its value to avoid exposing it to the user programatically (concern previouosly addressed in [SPARK-41442](https://issues.apache.org/jira/browse/SPARK-41442)).

For many SQLMetrics, we use initValue = -1 and zeroValue = 0 to indicate that the metric is by default invalid. Whenever an invalid metric is updated, it sets itself to zeroValue and becomes valid. Invalid metrics will be filtered out when calculating min, max, etc. as a workaround for [SPARK-11013](https://issues.apache.org/jira/browse/SPARK-11013).

### Why are the changes needed?

The semantics of initValue and _zeroValue in SQLMetrics is a little bit confusing, since they effectively mean the same thing. Changing it to the following would be clearer, especially in terms of defining what an "invalid" metric is.

### Does this PR introduce _any_ user-facing change?
No. This shouldn't change any behavior.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #44222 from davintjong-db/sqlmetric-initvalue-refactor.

Authored-by: Davin Tjong <davin.tjong@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>