[MINOR][DOCS] Fix rst link in Python API docs for .sql()
### What changes were proposed in this pull request?

This PR fixes the rst markup for a link in the documentation for `pyspark.sql.SparkSession.sql` and `pyspark.pandas.sql`.

### Why are the changes needed?

The current markup is incorrect.

Technically, though the markup in this PR is correct, the link target is incorrect. We should be linking to page relative to the site root, rather than hardcoding a link to `/latest/`. However, I could not figure out how to do that in rst, and building the API docs takes a really long time, and I could not make it build incrementally.

### Does this PR introduce _any_ user-facing change?

Yes, the markup goes from looking like this:

<img width="500" src="https://github.com/apache/spark/assets/1039369/077566a3-79df-4aa2-a0f7-d819f608f673">

To looking like this:

<img width="500" src="https://github.com/apache/spark/assets/1039369/b1453761-3f9c-435e-89e1-cfd3748cce9c">

### How was this patch tested?

I built the docs as follows:

```
SKIP_SCALADOC=1 SKIP_RDOC=1 SKIP_SQLDOC=1 bundle exec jekyll serve
```

And reviewed the output in my browser.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44488 from nchammas/data-types-rst-link.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>