[SPARK-46502][SQL] Support timestamp types in UnwrapCastInBinaryComparison
### What changes were proposed in this pull request?

This patch adds one more case to the optimization rule `UnwrapCastInBinaryComparison`. The added case handles unwrapping cast in binary comparison between timestamp and timestamp_ntz types.

### Why are the changes needed?

We encountered a case that a predicate on a timestamp_ntz column is not pushed down. The query plan looks like:

```
== Analyzed Logical Plan ==
batch: timestamp
Project [batch#26466]
+- Filter (batch#26466 >= cast(2023-12-21 10:00:00 as timestamp))
   +- SubqueryAlias spark_catalog.default.timestamp_view
      +- View (`spark_catalog`.`default`.`timestamp_view`, [batch#26466])
         +- Project [cast(batch#26467 as timestamp) AS batch#26466]
            +- Project [cast(batch#26463 as timestamp) AS batch#26467]
               +- SubqueryAlias spark_catalog.default.table_timestamp
                  +- Relation spark_catalog.default.table_timestamp[batch#26463] parquet

== Optimized Logical Plan ==
Project [cast(batch#26463 as timestamp) AS batch#26466]
+- Filter (isnotnull(batch#26463) AND (cast(batch#26463 as timestamp) >= 2023-12-21 10:00:00))
   +- Relation spark_catalog.default.table_timestamp[batch#26463] parquet
```

The predicate compares a timestamp_ntz column with a literal value. As the column is wrapped in a cast expression to timestamp type, the literal (string) is wrapped with a cast to timestamp type. The literal with cast is foldable so it is evaluated to literal of timestamp earlier. So the predicate becomes `cast(batch#26463 as timestamp) >= 2023-12-21 10:00:00`. As the cast is in column side, it cannot be pushed down to data source/table.

We have an optimization rule `UnwrapCastInBinaryComparison` that handles similar cases but it doesn't cover timestamp types.

After this patch, the query plan looks like:

```
== Analyzed Logical Plan ==
batch: timestamp
Project [batch#26446]
+- Filter (batch#26446 >= cast(2023-12-21 10:00:00 as timestamp))
   +- SubqueryAlias spark_catalog.default.timestamp_view
      +- View (`spark_catalog`.`default`.`timestamp_view`, [batch#26446])
         +- Project [cast(batch#26447 as timestamp) AS batch#26446]
            +- Project [cast(batch#26443 as timestamp) AS batch#26447]
               +- SubqueryAlias spark_catalog.default.table_timestamp
                  +- Relation spark_catalog.default.table_timestamp[batch#26443] parquet

== Optimized Logical Plan ==
Project [cast(batch#26443 as timestamp) AS batch#26446]
+- Filter (isnotnull(batch#26443) AND (batch#26443 >= 2023-12-21 10:00:00))
   +- Relation spark_catalog.default.table_timestamp[batch#26443] parquet
```

### Does this PR introduce _any_ user-facing change?

Yes. Casts between timestamp and timestamp_ntz in binary comparison will be unwrap if possible.

### How was this patch tested?

Added unit test.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #44480 from viirya/uncast.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>