[SPARK-46570][INFRA] Run Python 3.11 and 3.12 tests independently
### What changes were proposed in this pull request?

This PR aims to run `Python 3.11` and `3.12` tests independently.

### Why are the changes needed?

Currently, the CI is broken. It's easier to split the versions to isolate and identify the Python version-related root cause.

![Screenshot 2024-01-02 at 12 44 29 PM](https://github.com/apache/spark/assets/9700541/f034dc60-ae45-4538-8c2a-55bc89904d1f)

Since Python 3.12 is relatively new than the other versions, it needs more attention.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This should be tested after merging because `Python CI` is a daily test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44566 from dongjoon-hyun/SPARK-46570.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>