[SPARK-46585][CORE] Directly constructed `metricPeaks` as an `immutable.ArraySeq` instead of use `mutable.ArraySeq.toSeq` in `Executor`
### What changes were proposed in this pull request?
The following code exists in branch-3.5:

https://github.com/apache/spark/blob/fb90ade2c7390077d2755fc43b73e63f5cf44f21/core/src/main/scala/org/apache/spark/executor/Executor.scala#L744-L749

From the code comments, when using Scala 2.12, `metricsPoller.getTaskMetricPeaks(taskId)` is wrapped as `WrappedArray` to avoid collection copying, and the subsequent `metricPeaks.toSeq` is a redundant collection conversion for Scala 2.12.

However, for Scala 2.13, if `metricsPoller.getTaskMetricPeaks(taskId)` is still wrapped as `WrappedArray/mutable.ArraySeq`, it is impossible to avoid collection conversion, because the subsequent `metricPeaks.toSeq` will trigger a collection copy when using Scala 2.13.

So this pr changes the process to directly wrap `metricPeaks` as `immutable.ArraySeq` to ensure the same effect as when using Scala 2.12, which is also more in line with the original comment description.

### Why are the changes needed?
Avoid unnecessary collection copying when using Scala 2.13.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GitHub Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #44586 from LuciferYang/executor-metricPeaks.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
Co-authored-by: YangJie <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>