[SPARK-46568][PYTHON] Make Python data source options a case-insensitive dictionary
### What changes were proposed in this pull request?

This PR updates the `options` field to use a case-insensitive dictionary to keep the behavior consistent with the Scala side (which uses `CaseInsensitiveStringMap`). Currently, `options` are stored in a normal Python dictionary which can be confusing to users. For instance:
```python
class MyDataSource(DataSource):
    def __init__(self, options):
        self.api_key = options.get("API_KEY") # <- This is None

spark.read.format(..).option("API_KEY", my_key).load(...)
```
Here, `options` will not have this "API_KEY" as everything is converted to lowercase on the Scala side.

### Why are the changes needed?

To improve usability.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New unit tests

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #44564 from allisonwang-db/spark-46568-ds-options.

Authored-by: allisonwang-db <allison.wang@databricks.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>