[SPARK-46627][SS][UI] Fix timeline tooltip content on streaming ui
### What changes were proposed in this pull request?

Fix timeline tooltip content on streaming ui

### Why are the changes needed?

d3 v7 mouseover event has changed, we shall get the values from the 2nd parameter instead of the 1st one which points to the mouse event.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

![image](https://github.com/apache/spark/assets/8326978/ded6e09e-837d-499c-8fed-21e3f42d8244)

Additionally, https://issues.apache.org/jira/browse/SPARK-46631 was created to add UT for drawTimeline. We don't do it in this patch because it needs code refactoring beyond the scope of the current PR.

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #44633 from yaooqinn/SPARK-46627.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kousuke Saruta <sarutak@apache.org>