[SPARK-46633][SQL] Fix Avro reader to handle zero-length blocks
### What changes were proposed in this pull request?

This PR fixes a bug in Avro connector with regard to zero-length blocks. If a file contains one of these blocks, the Avro connector may return an incorrect number of records or even an empty DataFrame in some cases.

This was due to the way the `hasNextRow` check worked. `hasNext` method in Avro loads the next block so if the block is empty, it would return false and Avro connector will stop reading rows. However, we should continue checking the next block instead until the sync point.

### Why are the changes needed?

Fixes a correctness bug in Avro connector.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

I added a unit test and a generated sample file to verify the fix. Without the patch, reading such file would return fewer records or 0 compared to the actual number (depends on the maxPartitionBytes config).

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44635 from sadikovi/SPARK-46633.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>