[SPARK-46660][CONNECT] ReattachExecute requests updates aliveness of SessionHolder
### What changes were proposed in this pull request?

This PR makes `SparkConnectReattachExecuteHandler` fetch the `ExecuteHolder` via the`SessionHolder` which in turn refreshes it's aliveness. Further, this makes it consistent with `SparkConnectReleaseExecuteHandler`.

### Why are the changes needed?

Currently in ReattachExecute, we fetch the ExecuteHolder directly without going through the SessionHolder and hence the "aliveness" of the SessionHolder is not refreshed.

This would result in long-running queries (which do not send `ReleaseExecute` requests in specific) failing because the `SessionHolder` would expire from the cache during an active query execution.

### Does this PR introduce _any_ user-facing change?

Yes, it fixes the bug where long-running may fail when their corresponding `SessionHolder` is expired during active query execution.

### How was this patch tested?

New unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44670 from vicennial/SPARK-46660.

Authored-by: vicennial <venkata.gudesa@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>