[SPARK-46722][CONNECT] Add a test regarding to backward compatibility check for StreamingQueryListener in Spark Connect (Scala/PySpark)
### What changes were proposed in this pull request?

This PR proposes to add a functionality to perform backward compatibility check for StreamingQueryListener in Spark Connect (both Scala and PySpark), specifically implementing onQueryIdle or not.

### Why are the changes needed?

We missed to add backward compatibility test when introducing onQueryIdle, and it led to an issue in PySpark (https://issues.apache.org/jira/browse/SPARK-45631). We added the compatibility test in PySpark but didn't add it in Spark Connect.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Modified UTs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44736 from HeartSaVioR/SPARK-46722.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>